requirement.txt
Django==2.1.3
django-cors-headers==2.4.0
django-suit==0.2.25
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
pandas==0.24.1
xlrd==1.2.0
pillow==1.1.7
